+++ demo-output/src/main/java/io/pgenie/artifacts/myspace/musiccatalogue/statements/InsertAlbum.java
@@ -7,9 +7,8 @@ import java.sql.Date;
 import java.sql.Types;
 import java.time.*;
 import java.util.List;
-import io.codemine.java.postgresql.codecs.Codec;
-import io.pgenie.artifacts.myspace.musiccatalogue.JdbcCodec;
-import io.pgenie.artifacts.myspace.musiccatalogue.Statement;
+import io.codemine.java.postgresql.jdbc.Codec;
+import io.codemine.java.postgresql.jdbc.Statement;
 import io.pgenie.artifacts.myspace.musiccatalogue.types.*;
 
 /**
@@ -76,8 +75,8 @@ public record InsertAlbum(
     public void bindParams(PreparedStatement ps) throws SQLException {
         ps.setString(1, this.name());
         ps.setDate(2, Date.valueOf(this.released()));
-        new JdbcCodec<>(AlbumFormat.CODEC).bind(ps, 3, this.format());
-        new JdbcCodec<>(RecordingInfo.CODEC).bind(ps, 4, this.recording());
+        AlbumFormat.CODEC.bind(ps, 3, this.format());
+        RecordingInfo.CODEC.bind(ps, 4, this.recording());
     }
 
     @Override
@@ -89,7 +88,7 @@ public record InsertAlbum(
     public Result decodeResultSet(ResultSet rs) throws SQLException {
         rs.next();
 
-        long idCol = rs.getLong(1);
+        long idCol = Codec.INT8.decodeNonNullable(rs, 0, 1);
 
         return new Result(idCol);
     }