+++ demo-output/src/main/java/io/pgenie/artifacts/myspace/musiccatalogue/statements/SelectAlbumById.java
@@ -8,8 +8,8 @@ import java.sql.Types;
 import java.time.*;
 import java.util.List;
 import java.util.Optional;
-import io.pgenie.artifacts.myspace.musiccatalogue.JdbcCodec;
-import io.pgenie.artifacts.myspace.musiccatalogue.Statement;
+import io.codemine.java.postgresql.jdbc.Codec;
+import io.codemine.java.postgresql.jdbc.Statement;
 import io.pgenie.artifacts.myspace.musiccatalogue.types.*;
 
 /**
@@ -107,21 +107,13 @@ public record SelectAlbumById(
         if (!rs.next()) {
             return Optional.empty();
         }
-        long idCol = rs.getLong(1);
-        String nameCol = rs.getString(2);
-        Optional<LocalDate> releasedCol;
-        {
-            Date releasedColBase = rs.getDate(3);
-            if (releasedColBase != null) {
-                releasedCol = Optional.of(releasedColBase.toLocalDate());
-            } else {
-                releasedCol = Optional.empty();
-            }
-        }
-        Optional<AlbumFormat> formatCol = Optional.ofNullable(new JdbcCodec<>(AlbumFormat.CODEC).decodeNullable(rs, 0, 4));
-        Optional<RecordingInfo> recordingCol = Optional.ofNullable(new JdbcCodec<>(RecordingInfo.CODEC).decodeNullable(rs, 0, 5));
-        Optional<List<Optional<TrackInfo>>> tracksCol = Optional.ofNullable(new JdbcCodec<>(TrackInfo.CODEC.inDim()).decodeNullable(rs, 0, 6).stream().map(Optional::ofNullable).toList());
-        Optional<DiscInfo> discCol = Optional.ofNullable(new JdbcCodec<>(DiscInfo.CODEC).decodeNullable(rs, 0, 7));
+        long idCol = Codec.INT8.decodeNonNullable(rs, 0, 1);
+        String nameCol = Codec.TEXT.decodeNonNullable(rs, 0, 2);
+        Optional<LocalDate> releasedCol = Codec.DATE.decodeOptional(rs, 0, 3);
+        Optional<AlbumFormat> formatCol = AlbumFormat.CODEC.decodeOptional(rs, 0, 4);
+        Optional<RecordingInfo> recordingCol = RecordingInfo.CODEC.decodeOptional(rs, 0, 5);
+        Optional<List<Optional<TrackInfo>>> tracksCol = TrackInfo.CODEC.inDim().decodeOptional(rs, 0, 6).map(list1 -> list1.stream().map(Optional::ofNullable).toList());
+        Optional<DiscInfo> discCol = DiscInfo.CODEC.decodeOptional(rs, 0, 7);
 
         return Optional.of(new ResultRow(idCol, nameCol, releasedCol, formatCol, recordingCol, tracksCol, discCol));
     }